requirements.txt
boto3==1.42.15
botocore==1.42.15
fastapi==0.128.0
huggingface_hub==0.36.0
numpy==2.4.1
psycopg2_binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
PyYAML==6.0.3
scipy==1.17.0
torch==2.9.1
transformers==4.57.3
uvicorn>=0.38.0
gguf>=0.10.0
uvloop>=0.22.1
llama-cpp>=0.3.16